Roll forward: Make R8 warnings fatal in CaptivePortalLogin
Roll forward the change after dependencies on the Kotlin classes
that failed the build last time were removed.

This reverts commit 3ff76bda1770301b69e898d051a0c3436812da1b.

Ignoring R8 warnings is dangerous as it may change behavior when
optimizing; for example considering that code is unreachable when it
cannot find a classpath class.

Make the optimizer warnings fatal with ignore_warnings: false in
"optimize". This is a new property added to soong in a recent change.

Bug: 226127213
Change-Id: Ie29c4f20e8118cd884d2837b8305875c31f986a6
Test: m