Remove DISMISS_PORTAL_IN_VALIDATED_NETWORK flag
The flag was introduced to allow toggling NetworkMonitor-based
validation in CaptivePortalLoginActivity, but was never pushed, and the
default (enabled on R+) has always been used instead.

It can't be depended on to toggle behavior at runtime either, as the
flag would be cleared every time the device syncs its flags with the
server.

Remove the flag as pushing it at this point would not be worth the
effort, considering that only Q devices do not use the updated behavior.
Rolling out the feature on Q would be risky and have low value, and the
feature has been in use for long enough on R+ so that an emergency
disable will not be necessary.

Bug: 243315116
Test: atest
Change-Id: I844210de62f9d87316379cee462a9fa5a079518b