Do not use mock(WifiManager.class) in tests
mock(WifiManager.class) will throw a ClassNotFoundException on R when a
newer wifi module is installed, because newer methods were added that
reference classes that do not exist on R.

This means that as it is, it is no longer possible to mock WifiManager
with Mockito. Mock the only method usage using a different method.

Bug: 256774041
Test: atest
Change-Id: I0f86c45d3f25ac2df4ede70f8b86845e575f5eb9