Log deletes of downloaded files
There are some test flakes observing that files in the download
directory are deleted a few seconds after being created, during the
test. They do not indicate that CaptivePortalLoginActivity is started at
that time, but this is the only place where the files are deleted, so
add some logs to figure out if that is the culprit.

This path should only be used when passpoint files have been downloaded,
to clean them up in the next activity start.

Bug: 317602748
Test: atest
Change-Id: I38ed5ce62a708d6a30165151b83476e7e485eb40