Fix spinner hide testing in directly open download
Previously the spinner was not dismissed because of the
onDownloadComplete callback, because the downloadId was not mocked so
it didn't match the expected mDirectlyOpenId in the callback.

The test still passed on most devices because the intent triggering
the downloaded file install was not mocked either, so it would invoke
a system UI flow which puts the activity in the background and
dismiss the spinner. However on some devices that system flow is
just a toast and the test fails.

Add the missing mocking so the test covers what it is supposed to.

Bug: 314027069
Test: atest
Change-Id: Ieee581bb7c3c72d44071af21d26256306982b898