Remove black page when exiting the custom tab
When the custom tab is closed or minimized, the activity behind it
is revealed. In this case, this is the empty activity of the captive
portal login app, which serves no purpose. It should just be finished
as soon as the custom tab is launched so the activity behind it is
revealed instead.

See the bug for videos of what it looks like before/after.

Test: manual
Bug: 394720873
Change-Id: I19f280b5a7743e8b65dc9f904e89661b344fec3f